Add a stub for the mremap syscall
Previously, the mremap syscall was not being handled and was thus
crashing the VM. However, this syscall is sometimes used when using
realloc on large allocations in musl libc. To prevent the crash, a stub
implementation has been added which just returns -1. This indicates to
musl libc that an error occurred and it will then proceed to use a
fallback path to perform its realloc. While not great for performance,
this does exhibit correct behaviour.